refactor: 기존 util 분리에 의해 message generator 도 분리
prettier 적용과 userInfoType에 버전 설명도 추가하였습니다.

- close #190
