+++ src/message-generator/message-generator.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { MessageGenerator } from './message-generator.component';
+
+@Module({
+  providers: [MessageGenerator],
+  exports: [MessageGenerator],
+})
+export class MessageGeneratorModule {}
