+++ src/tag-log/dto/info-message.dto.ts
@@ -1,10 +1,9 @@
-
 export class InfoMessageDto {
-    title: string | null;
-    content: string | null;
+  title: string | null;
+  content: string | null;
 
-    constructor(title: string | null, content: string | null) {
-        this.title = title;
-        this.content = content;
-    }
+  constructor(title: string | null, content: string | null) {
+    this.title = title;
+    this.content = content;
+  }
 }
